README: Refactor cram invocation (snippets in .readme.md)
Inspired by what I already have in strava-gear.

Running cram on the entire README.md means that any line indented with 4
spaces is interpreted by cram, which is especially troublesome in code
blocks that are already indented because they're in a bullet list. As a
workaround, I've been using Unicode zero-width spaces, but that breaks
copy-pasting from the README.

Be explicit about what sections are fed to cram.
Possibly cleaner in general.